cr: reduce some code duplication, some comments
